Improve the HDF5 performance of the log file.
JSON + older HDF5 should be deprecated at some point.
