+++ tests/state_files/test_1_inst.json
@@ -1,7 +1,7 @@
 {
   "time": {
     "start_time": "2020-01-09 16:10:26.263195",
-    "end_time": "2020-03-26 13:29:47.470026"
+    "end_time": "2020-04-15 21:03:12.450837"
   },
   "version": "2.1",
   "software_version": "0.3.1+14.gd39b057.dirty",
